Revert "Merge pull request #64 from SheCodesAus/feature/search_functionality_BS"
This reverts commit 38729def3f008b63b26afdcf89188c3e766b5203, reversing
changes made to 5c1aac420dad1bf74b931c1bd73f99bb21c94a74.
