+++ src/pages/HomePage.jsx
@@ -1,7 +1,6 @@
 import BackgroundSlideshow from "../components/BackgroundSlideshow";
-import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-
+import React, { useState } from "react";
+import { useNavigate, Link } from "react-router-dom";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
 import "./HomePage.css";
@@ -10,38 +9,19 @@ import "../styles.css";
 import useRolemodels from "../hooks/use-rolemodels";
 import RolemodelCard from "../components/RolemodelCard";
 
-import SearchBar from "../components/SearchBar";
-
 function HomePage() {
     const navigate = useNavigate();
-    const [filteredRolemodels, setFilteredRolemodels] = useState([]);
+
     const { rolemodels, isLoading, error } = useRolemodels();
+    const [selectedLocation, setSelectedLocation] = useState("");
 
-
+    const handleSearch = () => {
+        console.log(`Searching for role models in: ${selectedLocation}`);
+        // Add functionality to filter role models based on location
+    };
 
     const handleSignUpNavigation = (type) => {
         navigate(`/signup?type=${type}`);
-    };
-
-    useEffect(() => {
-        // Set default grid with all role models when data is loaded
-        setFilteredRolemodels(rolemodels);
-    }, [rolemodels]);
-
-    const handleFilter = (selectedLocation) => {
-        if (!selectedLocation) {
-            // If no location is selected, show all role models
-            setFilteredRolemodels(rolemodels);
-        } else {
-            // Filter based on the selected location
-            const filtered = rolemodels.filter(
-                (rolemodel) =>
-                    rolemodel.location &&
-                    rolemodel.location.toLowerCase() ===
-                    selectedLocation.toLowerCase()
-            );
-            setFilteredRolemodels(filtered);
-        }
     };
 
     if (isLoading) {
@@ -51,7 +31,6 @@ function HomePage() {
     if (error) {
         return (<p>Error: {error.message}</p>);
     }
-
 
 
     return (
@@ -68,6 +47,8 @@ function HomePage() {
             <div id="app" className="featured_rolemodels">
                 <div className="sheinspires_header">
                     <div className="sheinspires_header_description">
+                        {/* <h1>She Inspires!</h1>
+                        <p>nice content about website in 1-2 sentence</p> */}
                     </div>
                 </div>
             </div>
@@ -76,10 +57,21 @@ function HomePage() {
             {/* Search  bar */}
 
             <div className="search-section">
-                <SearchBar onFilter={handleFilter} />
+                <select
+                    value={selectedLocation}
+                    onChange={(e) => setSelectedLocation(e.target.value)}
+                    className="location-dropdown"
+                >
+                    <option value="">Select a Location</option>
+                    <option value="Perth">Perth</option>
+                    <option value="Sydney">Sydney</option>
+                    <option value="Melbourne">Melbourne</option>
+                    <option value="Brisbane">Brisbane</option>
+                </select>
+                <button onClick={handleSearch} className="search-button">
+                    Search
+                </button>
             </div>
-
-
 
             {/* sign_buttons (she inspires, to be inspired , ) */}
 
@@ -109,17 +101,13 @@ function HomePage() {
 
             <div id="rolemodel-list">
                 <h1> Featured Role Models</h1>
-                <div className="cards-container">
-                    {filteredRolemodels.length > 0 ? (
-                        filteredRolemodels.map((rolemodelData, key) => (
-                            <RolemodelCard key={key} rolemodelData={rolemodelData} />
-                        ))
-                    ) : (
-                        <p>No role models found for the selected location.</p>
-                    )}
+                <div class="cards-container">
+
+                    {rolemodels.map((rolemodelData, key) => (
+                        <RolemodelCard key={key} rolemodelData={rolemodelData} />
+                    ))}
                 </div>
             </div>
-
